part_3/src/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
djangorestframework==3.16.0
et_xmlfile==2.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
psycopg==3.2.6
psycopg2==2.9.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
